Refactor: move latencyprofileobserver supressors to separate file
